chore: migrated test functions to testhelpers dir
